We need to finish building launchd with -Wshorten-64-to-32 someday...
git-svn-id: http://svn.macosforge.org/repository/launchd/trunk@23388 e92bca22-270c-0410-9cea-e3f1106b6a1c

+++ launchd/src/launchproxy.c
@@ -108,7 +108,7 @@ int main(int argc __attribute__((unused)), char *argv[])
 
 	tmp = launch_data_dict_lookup(resp, LAUNCH_JOBKEY_TIMEOUT);
 	if (tmp)
-		timeout.tv_sec = launch_data_get_integer(tmp);
+		timeout.tv_sec = (int)launch_data_get_integer(tmp);
 
 	tmp = launch_data_dict_lookup(resp, LAUNCH_JOBKEY_PROGRAM);
 	if (tmp)
